Functionality moved to a static lib, error message added, still no full text-validation
